requirement.txt
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
python-dotenv==0.13.0
redis==3.5.3
requests==2.23.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.2
Werkzeug==1.0.1
wincertstore==0.2
